Convert README CoffeeScript to JavaScript
CoffeeScript can be difficult for beginners to translate to JavaScript
or understand the difference between the two. This is especially true
with frameworks like Phoenix, which uses brunch as for their front-end
assets throwing beginners directly into JavaScript.

It's also reasonable to assume that users of CoffeeScript will have a
much easier time converting from JavaScript to CoffeeScript.